Reorganized code and added __getattr__ to VariableKiosk
The code was reorganized to make it more readable because some of the
equations were not in a logical order. The latter is a result of FST
code as FST itself sorts the equations into the right order.
Further a __getattr__ method was added to the VariableKiosk class so
instead of:
DVS = self.kiosk["DVS"]
we can now write:
self.kiosk.DVS
The latter makes refactoring much easier.